feat: Implement academic management system with frontend and backend
- Added frontend components for report generation.
- Set up main entry point for React application.
- Created global styles for the application.
- Developed build and run script for the entire system.
- Configured CORS for backend API.
- Implemented ReporteIntegralEstudianteController for fetching student reports.
- Created DTOs for Curso, Estudiante, ProyectoInvestigacion, and ReporteIntegralEstudiante.
- Developed services for managing courses, students, and research projects.
- Integrated PostgreSQL, MySQL, and MongoDB for data management.
- Added script to start services with Docker.

+++ src/main/java/pe/edu/utp/gestionacademicautp/repository/postgres/EstudianteRepository.java
@@ -3,5 +3,5 @@ package pe.edu.utp.gestionacademicautp.repository.postgres;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.edu.utp.gestionacademicautp.model.postgres.Estudiante;
 
-public interface EstudianteRepository extends JpaRepository<Estudiante, Long> {
+public interface EstudianteRepository extends JpaRepository<Estudiante, Integer> {
 }
